Add files via upload
Gents,

I have fixed the issues with the UI and have added the toolbar and back button on top. once Guru works on the JAVA a bit I will also add the drawers at the bottom or on left of the screen.

Also, the home screen is now clickable instead of only the button.